feature/45 - allow multiple paths to be measured
1. Use Ctrl to start a new path
2. Start clicking (use shift to snap)
3. Release Ctrl to end that path
4. Go to #1
5. Press Esc to clear the tool
